double -> int algorihtme ok

+++ src/Constante.java
@@ -1,6 +1,6 @@
 public class Constante {
 
-    public static double debitTotal;
+    public static int debitTotal;
     public static double elevAmont;
     public static double coeffPertes = 0.5 * Math.pow(10,-5);
